move field tests to file
